feat: split mod
